Add triplea-java-library convention plugin
This creates a pre-compiled script plugin to apply common setup across all projects in the build.  Previously, this was done via an allprojects block.

This is a more idiomatic way to share configuration across multiple projects, and will enable future Gradle build acceleration features.  Though it adds additional lines of configuration to simple projects like /game-app/map-data vs. the previous setup, this is a worthwhile change.